moved RaceModel and RaceDetailModel to domain.models
