+++ app/src/main/java/com/example/f1calendarOP/Differ.kt
@@ -1,8 +1,9 @@
 package com.example.f1calendarOP
 
 import androidx.recyclerview.widget.DiffUtil
+import com.example.f1calendarOP.domain.models.Race
 
-    val raceDiffCallback = object : DiffUtil.ItemCallback<Race>(){
+val raceDiffCallback = object : DiffUtil.ItemCallback<Race>(){
         override fun areItemsTheSame(oldItem: Race, newItem: Race): Boolean {
             return oldItem.round == newItem.round
         }
