+++ app/src/main/java/com/example/f1calendarOP/RaceDetailViewHolder.kt
@@ -6,6 +6,7 @@ import com.bumptech.glide.Glide
 import com.example.f1calendarOP.databinding.ItemDetailCircuitBinding
 import com.example.f1calendarOP.databinding.ItemDetailHeaderBinding
 import com.example.f1calendarOP.databinding.ItemDetailSessionBinding
+import com.example.f1calendarOP.domain.models.RaceDetailModel
 
 
 sealed class RaceDetailViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
