+++ app/src/main/java/com/example/f1calendarOP/RaceDetailViewModel.kt
@@ -3,6 +3,8 @@ package com.example.f1calendarOP
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import com.example.f1calendarOP.domain.models.Race
+import com.example.f1calendarOP.domain.models.RaceDetailModel
 
 class RaceDetailViewModel(private val raceModel : Race) : ViewModel() {
