+++ app/src/main/res/navigation/nav_graph_application.xml
@@ -21,7 +21,7 @@
             android:label="DetailFragment" >
             <argument
                 android:name="raceInfo"
-                app:argType="com.example.f1calendarOP.Race" />
+                app:argType="com.example.f1calendarOP.domain.models.Race" />
         </fragment>
         <fragment
             android:id="@+id/raceListFragment"
